fix(xlsx): use current picture payload on export
